+++ odoo/local-src/speedy_views/models/__init__.py
@@ -9,5 +9,6 @@ from . import sale_order
 from . import stock_move
 from . import stock_move_operation_link
 from . import stock_picking
+from . import stock_quant
 from . import stock_quant_package
 from . import stock_pack_operation