+++ odoo/local-src/speedy_views/models/stock_move.py
@@ -10,3 +10,6 @@ class StockMove(models.Model):
     _inherit = 'stock.move'
 
     procurement_id = fields.Many2one(index=True)
+    split_from = fields.Many2one(index=True)
+    origin_returned_move_id = fields.Many2one(index=True)
+    inventory_id = fields.Many2one(index=True)